requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
torch>=1.9.0
torchvision>=0.10.0
rembg>=2.0.30
matplotlib>=3.4.3
pyrender>=0.1.45
trimesh>=3.9.35
scipy>=1.7.1
Pillow>=8.3.1
transformers>=4.11.0
diffusers>=0.11.1
accelerate>=0.12.0
onnxruntime>=1.8.0 